Разбивает главную страницу на компоненты
