--- project/src/components/Readme.md
@@ -1,3 +1,0 @@
-# components
-
-Компоненты приложения. Каждый компонент размещается в отдельной директории.
