+++ project/src/components/button-show-more/button-show-more.tsx
@@ -1,0 +1,9 @@
+function ButtonShowMore(): JSX.Element {
+  return (
+    <div className="catalog__more">
+      <button className="catalog__button" type="button">Show more</button>
+    </div>
+  );
+}
+
+export default ButtonShowMore;
