+++ project/src/components/controls/controls.tsx
@@ -1,0 +1,20 @@
+function Controls(): JSX.Element {
+  return (
+    <div className="film-card__buttons">
+      <button className="btn btn--play film-card__button" type="button">
+        <svg viewBox="0 0 19 19" width="19" height="19">
+          <use xlinkHref="/play-s"></use>
+        </svg>
+        <span>Play</span>
+      </button>
+      <button className="btn btn--list film-card__button" type="button">
+        <svg viewBox="0 0 19 20" width="19" height="20">
+          <use xlinkHref="/add"></use>
+        </svg>
+        <span>My list</span>
+      </button>
+    </div>
+  );
+}
+
+export default Controls;
