+++ project/src/pages/main-page/main-page.tsx
@@ -1,0 +1,63 @@
+import Header from '../../components/header/header';
+import Controls from '../../components/controls/controls';
+import GenreList from '../../components/genre-list/genre-list';
+import FilmsList from '../../components/films-list/films-list';
+import ButtonShowMore from '../../components/button-show-more/button-show-more';
+import Footer from '../../components/footer/footer';
+import {Movies} from '../../types/movie';
+
+type MainPageProps = {
+  promoFilmName: string;
+  promoFilmGenre: string;
+  promoFilmYear: number;
+  films: Movies;
+}
+
+function MainPage({promoFilmName, promoFilmGenre, promoFilmYear, films}: MainPageProps): JSX.Element {
+  return (
+    <>
+      <section className="film-card">
+        <div className="film-card__bg">
+          <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
+        </div>
+
+        <h1 className="visually-hidden">WTW</h1>
+
+        <Header />
+
+        <div className="film-card__wrap">
+          <div className="film-card__info">
+            <div className="film-card__poster">
+              <img src="img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218" height="327" />
+            </div>
+
+            <div className="film-card__desc">
+              <h2 className="film-card__title">{promoFilmName}</h2>
+              <p className="film-card__meta">
+                <span className="film-card__genre">{promoFilmGenre}</span>
+                <span className="film-card__year">{promoFilmYear}</span>
+              </p>
+
+              <Controls />
+            </div>
+          </div>
+        </div>
+      </section>
+
+      <div className="page-content">
+        <section className="catalog">
+          <h2 className="catalog__title visually-hidden">Catalog</h2>
+
+          <GenreList />
+          <FilmsList films={films}/>
+          <ButtonShowMore />
+
+        </section>
+
+        <Footer />
+      </div>
+    </>
+  );
+}
+
+export default MainPage;
