+++ project/src/pages/my-list/my-list.tsx
@@ -1,3 +1,5 @@
+import Footer from '../../components/footer/footer';
+
 function MyList(): JSX.Element {
   return (
     <div className="user-page">
@@ -113,19 +115,8 @@ function MyList(): JSX.Element {
         </div>
       </section>
 
-      <footer className="page-footer">
-        <div className="logo">
-          <a href="main.html" className="logo__link logo__link--light">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </a>
-        </div>
+      <Footer />
 
-        <div className="copyright">
-          <p>© 2019 What to watch Ltd.</p>
-        </div>
-      </footer>
     </div>
   );
 }
