+++ project/src/pages/player/player.tsx
