+++ project/src/pages/sign-in/sign-in.tsx
@@ -1,3 +1,5 @@
+import Footer from '../../components/footer/footer';
+
 function SignIn(): JSX.Element {
   return (
     <div className="user-page">
@@ -31,19 +33,7 @@ function SignIn(): JSX.Element {
         </form>
       </div>
 
-      <footer className="page-footer">
-        <div className="logo">
-          <a href="main.html" className="logo__link logo__link--light">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </a>
-        </div>
-
-        <div className="copyright">
-          <p>© 2019 What to watch Ltd.</p>
-        </div>
-      </footer>
+      <Footer />
     </div>
   );
 }
