--- project/src/types/Readme.md
@@ -1,3 +1,0 @@
-# types
-
-Описание типов, интерфейсов.
