+++ project/src/types/movie.tsx
@@ -1,0 +1,7 @@
+export type Movie = {
+  id: number;
+  name: string;
+  previewImage: string;
+}
+
+export type Movies = Movie[];
